syntax: keep the offset of a '+' match in off_t
match_plus() remembers the byte the pattern last matched at and goes back to
it. That offset was kept in an int, so on a file over 2 GB it came back
negative and the match landed anywhere.

Only the type changes. Below 2 GB the two are the same, which is why no test
sees it.

+++ src/syntax/syntax.c
@@ -538,7 +538,7 @@ match_plus (const syntax_scanner_t *sc, const GString *text, const char *whole_r
 {
     const unsigned char *p = *pp + 1;
     off_t i = *ii;
-    int j = 0;
+    off_t j = 0;
 
     while (TRUE)
     {
